+++ src/utils/crop.js
@@ -10,18 +10,14 @@ function(
 	 * Crop - this class is used for selection crop above the image/element.
 	 * 
 	 * @param {Object} [options] Configuration
-	 * @param {Number} [options.canWidth] Canvas width
-	 * @param {Number} [options.canHeight] Canvas height
-	 * @param {Number} [options.zoom = 100] start zoom in [%]
-	 * @param {Number} [options.minZoom = 20] min zoom in [%]
-	 * @param {Number} [options.maxZoom = 100] max zoom in [%]
-	 * @param {Number} [options.zoomStep = 10] How many [%] add/dec with zoom change
-	 * @param {Number} [options.zoomMoveStep = 1] Under 100% multiplier for faster image movement
-	 * @param {Object} [options.curEntity = $anonymizer.ENTITES.CIRCLE] Start entity from $anonymizer.ENTITES
-	 * @param {Number} [options.showPreview = true] Show preview - image overview
-	 * @param {Number} [options.previewLeft = 17] Preview location from left top corner, axe x [px]
-	 * @param {Number} [options.previewTop = 17] Preview location from left top corner, axe y [px]
-	 * @param {Number} [options.previewWidth = 200] Preview image width [px]
+	 * @param {Number} [options.width = 250] Crop width
+	 * @param {Number} [options.height = 250] Crop height
+	 * @param {Number} [options.minWidth = 10] Crop min width, always higher than 0!
+	 * @param {Number} [options.minHeight = 10] Crop min height, always higher than 0!
+	 * @param {Number} [options.maxWidth = Infinity] Crop max width
+	 * @param {Number} [options.maxHeight = Infinity] Crop max height
+	 * @param {Boolean} [options.resizable = true] Crop can be resizabled by points
+	 * @param {Number} [options.aspectRatio = 0] Crop aspect ratio for width / height
 	 * @class $crop
 	 */
 	var $crop = function(options) {
@@ -33,11 +29,11 @@ function(
 			width: 250, // initial size
 			height: 250,
 			minWidth: 10,
-			minHeight: 10, // always higher than 0! if resizable=true
+			minHeight: 10, //  if resizable=true
 			maxWidth: Infinity,
 			maxHeight: Infinity,
 			resizable: true,
-			aspectRatio: null
+			aspectRatio: 0
 		};
 
 		for (var op in options) {
@@ -82,6 +78,8 @@ function(
 		this._dom = {};
 
 		this._create();
+
+		this._dom.container.addEventListener("mousedown", this._binds.mouseDown);
 
 		// crop is by default hidden
 		this.hide();
@@ -141,8 +139,6 @@ function(
 			}],
 			_exported: "container"
 		}, this._dom);
-
-		this._dom.container.addEventListener("mousedown", this._binds.mouseDown);
 	};
 
 	/**
@@ -181,14 +177,16 @@ function(
 	 */
 	$crop.prototype._alignPoints = function() {
 		var p = this._points;
+		var w = this._dim.areaWidth - this._dim.width;
+		var h = this._dim.areaHeight - this._dim.height;
 
-		p.nw.x = $math.setRange(p.nw.x, 0, this._dim.areaWidth - this._dim.width);
-		p.sw.x = $math.setRange(p.sw.x, 0, this._dim.areaWidth - this._dim.width);
+		p.nw.x = $math.setRange(p.nw.x, 0, w);
+		p.sw.x = $math.setRange(p.sw.x, 0, w);
 		p.ne.x = $math.setRange(p.ne.x, this._dim.width, this._dim.areaWidth);
 		p.se.x = $math.setRange(p.se.x, this._dim.width, this._dim.areaWidth);
 
-		p.nw.y = $math.setRange(p.nw.y, 0, this._dim.areaHeight - this._dim.height);
-		p.ne.y = $math.setRange(p.ne.y, 0, this._dim.areaHeight - this._dim.height);
+		p.nw.y = $math.setRange(p.nw.y, 0, h);
+		p.ne.y = $math.setRange(p.ne.y, 0, h);
 		p.sw.y = $math.setRange(p.sw.y, this._dim.height, this._dim.areaHeight);
 		p.se.y = $math.setRange(p.se.y, this._dim.height, this._dim.areaHeight);
 	};
@@ -422,6 +420,41 @@ function(
 	};
 
 	/**
+	 * Fit crop to whole area and center him on the screen.
+	 * 
+	 * @member $crop
+	 */
+	$crop.prototype.fitToArea = function() {
+		var width;
+		var height;
+
+		if (this._options.aspectRatio) {
+			var ratio = this._options.aspectRatio;
+
+			// try width
+			width = this._dim.areaWidth;
+			height = Math.round(width / ratio);
+
+			// try height
+			if (height > this._dim.areaHeight) {
+				height = this._dim.areaHeight;
+				width = Math.round(height * ratio);
+			}
+		}
+		else {
+			width = Math.min(this._options.maxWidth, this._dim.areaWidth);
+			height = Math.min(this._options.maxHeight, this._dim.areaHeight);
+		}
+		
+		// update dimensions
+		this._dim.width = width;
+		this._dim.height = height;
+
+		// center and redraw
+		this.setCenter();
+	};
+
+	/**
 	 * Remove crop from DOM.
 	 * 
 	 * @member $crop
@@ -432,6 +465,8 @@ function(
 		if (c.parentNode) {
 			c.parentNode.removeChild(c);
 		}
+
+		this._dom.container.removeEventListener("mousedown", this._binds.mouseDown);
 	};
 
 	/**
@@ -445,38 +480,27 @@ function(
 	};
 
 	/**
-	 * Set crop dimensions.
+	 * Set crop area dimensions.
 	 * 
-	 * @param {Object} dim
-	 * @param {Number} dim.width Area width
-	 * @param {Number} dim.height Area height
+	 * @param {Object} [dim]
+	 * @param {Number} [dim.areaWidth] Area width
+	 * @param {Number} [dim.areaHeight] Area height
 	 * @member $crop
 	 */
 	$crop.prototype.setDim = function(dim) {
-		if (!dim) {
-			return;
+		dim = dim || {};
+
+		if (dim.areaWidth) {
+			this._dim.areaWidth = dim.areaWidth;
+
+			this._dom.container.style.width = this._dim.areaWidth + "px";
 		}
 
-		var areaWidth = dim.width;
-		var areaHeight = dim.height;
+		if (dim.areaHeight) {
+			this._dim.areaHeight = dim.areaHeight;
 
-		this._dim.areaWidth = areaWidth;
-		this._dim.areaHeight = areaHeight;
-
-		this._dom.container.style.width = areaWidth + "px";
-		this._dom.container.style.height = areaHeight + "px";
-
-		var width = Math.min(this._dim.width, this._dim.areaWidth);
-		var height = Math.min(this._dim.height, this._dim.areaHeight);
-
-		if (this._options.aspectRatio) {
-			height = Math.round(width / this._options.aspectRatio);
+			this._dom.container.style.height = this._dim.areaHeight + "px";
 		}
-
-		this._dim.width = width;
-		this._dim.height = height;
-
-		this.setCenter();
 	};
 
 	/**